working /v1/messages endpoint
